Update view structure
